test(mobile): split oversized test files to ≤500 LOC
- row-transforms-misc.test.ts (563 LOC) → row-transforms-documents (notes/journal/wiki) + row-transforms-lifecycle-fields (lifecycle/timer/checkin/field/friend)
- row-transforms-comms.test.ts (578 LOC) → row-transforms-channels (channel/message/boardMessage) + row-transforms-polls-social (poll/ack/relationship/privacy/friend-conn)
- SyncProvider-lifecycle.test.tsx (700 LOC) → split bootstrap-error nested describe into SyncProvider-bootstrap.test.tsx
- SyncProvider-pipeline.test.tsx (693 LOC) → split materializer-subscriber describe into SyncProvider-materializer.test.tsx
- opfs-worker-dispatch.test.ts (680 LOC) → split LRU/cache/error tests into opfs-worker-cache.test.ts
- trpc-persister-api-core.test.ts (693 LOC) → extract MockTRPCClient to helpers/trpc-mock-client.ts; split fronting tests into trpc-persister-api-fronting.test.ts
- trpc-persister-api-comms.test.ts → use shared mock helper; split group/blob/importEntityRef into trpc-persister-api-bulk.test.ts

All in-scope files now ≤500 LOC. 140 test files, 1366 tests pass.
