+++ apps/mobile/src/data/__tests__/row-transforms/row-transforms-structure-innerworld.test.ts
@@ -19,7 +19,7 @@ import {
   rowToStructureEntityLink,
   rowToStructureEntityMemberLink,
   rowToStructureEntityType,
-} from "../row-transforms/index.js";
+} from "../../row-transforms/index.js";
 
 // ── rowToStructureEntity ──────────────────────────────────────────────────────
