+++ apps/mobile/src/features/import-sp/__tests__/trpc-persister-api-core.test.ts
@@ -1,290 +1,21 @@
 /**
- * tRPC-backed PersisterApi bridge tests — core entity groups.
+ * tRPC PersisterApi bridge tests — system, settings, bucket, field, customFront,
+ * member, friend.
  *
- * Covers: system, systemSettings, bucket, field, customFront, member, friend,
- *         frontingSession, frontingComment, note
- * Companion file: trpc-persister-api-comms.test.ts
+ * Companion files: trpc-persister-api-fronting.test.ts,
+ *                  trpc-persister-api-comms.test.ts
  */
 
 import { brandId } from "@pluralscape/types";
-import { describe, expect, it, vi } from "vitest";
+import { describe, expect, it } from "vitest";
 
 import { createTRPCPersisterApi } from "../trpc-persister-api.js";
 
-import type { TRPCClientSubset } from "../trpc-persister-api.js";
-import type { SystemId } from "@pluralscape/types";
-import type { Mock } from "vitest";
+import { makeMockClient } from "./helpers/trpc-mock-client.js";
 
-// ── Helpers ─────────────────────────────────────────────────────────
+import type { SystemId } from "@pluralscape/types";
 
 const TEST_SYSTEM_ID = brandId<SystemId>("sys_bridge_test");
-
-interface MockQuery<TInput, TOutput> {
-  readonly query: Mock<(input: TInput) => Promise<TOutput>>;
-}
-
-interface MockMutation<TInput, TOutput> {
-  readonly mutate: Mock<(input: TInput) => Promise<TOutput>>;
-}
-
-interface MockTRPCClient extends TRPCClientSubset {
-  readonly system: {
-    readonly get: MockQuery<
-      Parameters<TRPCClientSubset["system"]["get"]["query"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["system"]["get"]["query"]>>
-    >;
-    readonly update: MockMutation<
-      Parameters<TRPCClientSubset["system"]["update"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["system"]["update"]["mutate"]>>
-    >;
-  };
-  readonly systemSettings: {
-    readonly settings: {
-      readonly get: MockQuery<
-        Parameters<TRPCClientSubset["systemSettings"]["settings"]["get"]["query"]>[0],
-        Awaited<ReturnType<TRPCClientSubset["systemSettings"]["settings"]["get"]["query"]>>
-      >;
-      readonly update: MockMutation<
-        Parameters<TRPCClientSubset["systemSettings"]["settings"]["update"]["mutate"]>[0],
-        Awaited<ReturnType<TRPCClientSubset["systemSettings"]["settings"]["update"]["mutate"]>>
-      >;
-    };
-  };
-  readonly bucket: {
-    readonly create: MockMutation<
-      Parameters<TRPCClientSubset["bucket"]["create"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["bucket"]["create"]["mutate"]>>
-    >;
-    readonly update: MockMutation<
-      Parameters<TRPCClientSubset["bucket"]["update"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["bucket"]["update"]["mutate"]>>
-    >;
-  };
-  readonly field: {
-    readonly definition: {
-      readonly create: MockMutation<
-        Parameters<TRPCClientSubset["field"]["definition"]["create"]["mutate"]>[0],
-        Awaited<ReturnType<TRPCClientSubset["field"]["definition"]["create"]["mutate"]>>
-      >;
-      readonly update: MockMutation<
-        Parameters<TRPCClientSubset["field"]["definition"]["update"]["mutate"]>[0],
-        Awaited<ReturnType<TRPCClientSubset["field"]["definition"]["update"]["mutate"]>>
-      >;
-    };
-    readonly value: {
-      readonly set: MockMutation<
-        Parameters<TRPCClientSubset["field"]["value"]["set"]["mutate"]>[0],
-        Awaited<ReturnType<TRPCClientSubset["field"]["value"]["set"]["mutate"]>>
-      >;
-    };
-  };
-  readonly customFront: {
-    readonly create: MockMutation<
-      Parameters<TRPCClientSubset["customFront"]["create"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["customFront"]["create"]["mutate"]>>
-    >;
-    readonly update: MockMutation<
-      Parameters<TRPCClientSubset["customFront"]["update"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["customFront"]["update"]["mutate"]>>
-    >;
-  };
-  readonly member: {
-    readonly create: MockMutation<
-      Parameters<TRPCClientSubset["member"]["create"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["member"]["create"]["mutate"]>>
-    >;
-    readonly update: MockMutation<
-      Parameters<TRPCClientSubset["member"]["update"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["member"]["update"]["mutate"]>>
-    >;
-  };
-  readonly frontingSession: {
-    readonly create: MockMutation<
-      Parameters<TRPCClientSubset["frontingSession"]["create"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["frontingSession"]["create"]["mutate"]>>
-    >;
-    readonly update: MockMutation<
-      Parameters<TRPCClientSubset["frontingSession"]["update"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["frontingSession"]["update"]["mutate"]>>
-    >;
-  };
-  readonly frontingComment: {
-    readonly create: MockMutation<
-      Parameters<TRPCClientSubset["frontingComment"]["create"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["frontingComment"]["create"]["mutate"]>>
-    >;
-    readonly update: MockMutation<
-      Parameters<TRPCClientSubset["frontingComment"]["update"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["frontingComment"]["update"]["mutate"]>>
-    >;
-  };
-  readonly note: {
-    readonly create: MockMutation<
-      Parameters<TRPCClientSubset["note"]["create"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["note"]["create"]["mutate"]>>
-    >;
-    readonly update: MockMutation<
-      Parameters<TRPCClientSubset["note"]["update"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["note"]["update"]["mutate"]>>
-    >;
-  };
-  readonly poll: {
-    readonly create: MockMutation<
-      Parameters<TRPCClientSubset["poll"]["create"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["poll"]["create"]["mutate"]>>
-    >;
-    readonly update: MockMutation<
-      Parameters<TRPCClientSubset["poll"]["update"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["poll"]["update"]["mutate"]>>
-    >;
-    readonly castVote: MockMutation<
-      Parameters<TRPCClientSubset["poll"]["castVote"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["poll"]["castVote"]["mutate"]>>
-    >;
-  };
-  readonly channel: {
-    readonly create: MockMutation<
-      Parameters<TRPCClientSubset["channel"]["create"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["channel"]["create"]["mutate"]>>
-    >;
-    readonly update: MockMutation<
-      Parameters<TRPCClientSubset["channel"]["update"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["channel"]["update"]["mutate"]>>
-    >;
-  };
-  readonly message: {
-    readonly create: MockMutation<
-      Parameters<TRPCClientSubset["message"]["create"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["message"]["create"]["mutate"]>>
-    >;
-    readonly update: MockMutation<
-      Parameters<TRPCClientSubset["message"]["update"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["message"]["update"]["mutate"]>>
-    >;
-  };
-  readonly boardMessage: {
-    readonly create: MockMutation<
-      Parameters<TRPCClientSubset["boardMessage"]["create"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["boardMessage"]["create"]["mutate"]>>
-    >;
-    readonly update: MockMutation<
-      Parameters<TRPCClientSubset["boardMessage"]["update"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["boardMessage"]["update"]["mutate"]>>
-    >;
-  };
-  readonly group: {
-    readonly create: MockMutation<
-      Parameters<TRPCClientSubset["group"]["create"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["group"]["create"]["mutate"]>>
-    >;
-    readonly update: MockMutation<
-      Parameters<TRPCClientSubset["group"]["update"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["group"]["update"]["mutate"]>>
-    >;
-    readonly addMember: MockMutation<
-      Parameters<TRPCClientSubset["group"]["addMember"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["group"]["addMember"]["mutate"]>>
-    >;
-  };
-  readonly blob: {
-    readonly createUploadUrl: MockMutation<
-      Parameters<TRPCClientSubset["blob"]["createUploadUrl"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["blob"]["createUploadUrl"]["mutate"]>>
-    >;
-    readonly confirmUpload: MockMutation<
-      Parameters<TRPCClientSubset["blob"]["confirmUpload"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["blob"]["confirmUpload"]["mutate"]>>
-    >;
-  };
-  readonly importEntityRef: {
-    readonly lookupBatch: MockMutation<
-      Parameters<TRPCClientSubset["importEntityRef"]["lookupBatch"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["importEntityRef"]["lookupBatch"]["mutate"]>>
-    >;
-    readonly upsertBatch: MockMutation<
-      Parameters<TRPCClientSubset["importEntityRef"]["upsertBatch"]["mutate"]>[0],
-      Awaited<ReturnType<TRPCClientSubset["importEntityRef"]["upsertBatch"]["mutate"]>>
-    >;
-  };
-}
-
-function makeMockClient(): MockTRPCClient {
-  return {
-    system: {
-      get: { query: vi.fn() },
-      update: { mutate: vi.fn() },
-    },
-    systemSettings: {
-      settings: {
-        get: { query: vi.fn() },
-        update: { mutate: vi.fn() },
-      },
-    },
-    bucket: {
-      create: { mutate: vi.fn() },
-      update: { mutate: vi.fn() },
-    },
-    field: {
-      definition: {
-        create: { mutate: vi.fn() },
-        update: { mutate: vi.fn() },
-      },
-      value: {
-        set: { mutate: vi.fn() },
-      },
-    },
-    customFront: {
-      create: { mutate: vi.fn() },
-      update: { mutate: vi.fn() },
-    },
-    member: {
-      create: { mutate: vi.fn() },
-      update: { mutate: vi.fn() },
-    },
-    frontingSession: {
-      create: { mutate: vi.fn() },
-      update: { mutate: vi.fn() },
-    },
-    frontingComment: {
-      create: { mutate: vi.fn() },
-      update: { mutate: vi.fn() },
-    },
-    note: {
-      create: { mutate: vi.fn() },
-      update: { mutate: vi.fn() },
-    },
-    poll: {
-      create: { mutate: vi.fn() },
-      update: { mutate: vi.fn() },
-      castVote: { mutate: vi.fn() },
-    },
-    channel: {
-      create: { mutate: vi.fn() },
-      update: { mutate: vi.fn() },
-    },
-    message: {
-      create: { mutate: vi.fn() },
-      update: { mutate: vi.fn() },
-    },
-    boardMessage: {
-      create: { mutate: vi.fn() },
-      update: { mutate: vi.fn() },
-    },
-    group: {
-      create: { mutate: vi.fn() },
-      update: { mutate: vi.fn() },
-      addMember: { mutate: vi.fn() },
-    },
-    blob: {
-      createUploadUrl: { mutate: vi.fn() },
-      confirmUpload: { mutate: vi.fn() },
-    },
-    importEntityRef: {
-      lookupBatch: { mutate: vi.fn() },
-      upsertBatch: { mutate: vi.fn() },
-    },
-  };
-}
 
 // ── system ──────────────────────────────────────────────────────────
 
@@ -555,139 +286,5 @@ describe("friend", () => {
 
     expect(a.placeholderId).toBe("import_friend_user_a");
     expect(b.placeholderId).toBe("import_friend_user_b");
-  });
-});
-
-// ── frontingSession ─────────────────────────────────────────────────
-
-describe("frontingSession", () => {
-  it("create calls frontingSession.create with startTime", async () => {
-    const client = makeMockClient();
-    client.frontingSession.create.mutate.mockResolvedValue({ id: "fs_1", version: 1 });
-    const api = createTRPCPersisterApi(client);
-
-    const result = await api.frontingSession.create(TEST_SYSTEM_ID, {
-      encryptedData: "enc_session",
-      startTime: 1_700_000_000,
-      memberId: null,
-      customFrontId: null,
-      structureEntityId: null,
-    });
-
-    expect(result).toEqual({ id: "fs_1", version: 1 });
-    expect(client.frontingSession.create.mutate).toHaveBeenCalledWith({
-      systemId: TEST_SYSTEM_ID,
-      encryptedData: "enc_session",
-      startTime: 1_700_000_000,
-      memberId: null,
-      customFrontId: null,
-      structureEntityId: null,
-    });
-  });
-
-  it("update calls frontingSession.update with sessionId", async () => {
-    const client = makeMockClient();
-    client.frontingSession.update.mutate.mockResolvedValue({ id: "fs_1", version: 2 });
-    const api = createTRPCPersisterApi(client);
-
-    const result = await api.frontingSession.update(TEST_SYSTEM_ID, "fs_1", {
-      encryptedData: "enc_session_v2",
-      version: 1,
-    });
-
-    expect(result).toEqual({ id: "fs_1", version: 2 });
-    expect(client.frontingSession.update.mutate).toHaveBeenCalledWith({
-      systemId: TEST_SYSTEM_ID,
-      sessionId: "fs_1",
-      encryptedData: "enc_session_v2",
-      version: 1,
-    });
-  });
-});
-
-// ── frontingComment ─────────────────────────────────────────────────
-
-describe("frontingComment", () => {
-  it("create calls frontingComment.create", async () => {
-    const client = makeMockClient();
-    client.frontingComment.create.mutate.mockResolvedValue({ id: "fcom_1", version: 1 });
-    const api = createTRPCPersisterApi(client);
-
-    const result = await api.frontingComment.create(TEST_SYSTEM_ID, {
-      encryptedData: "enc_comment",
-      sessionId: "fs_1",
-      memberId: null,
-      customFrontId: null,
-      structureEntityId: null,
-    });
-
-    expect(result).toEqual({ id: "fcom_1", version: 1 });
-    expect(client.frontingComment.create.mutate).toHaveBeenCalledWith({
-      systemId: TEST_SYSTEM_ID,
-      sessionId: "fs_1",
-      encryptedData: "enc_comment",
-    });
-  });
-
-  it("update calls frontingComment.update with commentId and sessionId", async () => {
-    const client = makeMockClient();
-    client.frontingComment.update.mutate.mockResolvedValue({ id: "fcom_1", version: 2 });
-    const api = createTRPCPersisterApi(client);
-
-    const result = await api.frontingComment.update(TEST_SYSTEM_ID, "fcom_1", {
-      encryptedData: "enc_comment_v2",
-      version: 1,
-      sessionId: "fs_real_session",
-    });
-
-    expect(result).toEqual({ id: "fcom_1", version: 2 });
-    expect(client.frontingComment.update.mutate).toHaveBeenCalledWith({
-      systemId: TEST_SYSTEM_ID,
-      sessionId: "fs_real_session",
-      commentId: "fcom_1",
-      encryptedData: "enc_comment_v2",
-      version: 1,
-    });
-  });
-});
-
-// ── note ────────────────────────────────────────────────────────────
-
-describe("note", () => {
-  it("create calls note.create", async () => {
-    const client = makeMockClient();
-    client.note.create.mutate.mockResolvedValue({ id: "note_1", version: 1 });
-    const api = createTRPCPersisterApi(client);
-
-    const result = await api.note.create(TEST_SYSTEM_ID, {
-      encryptedData: "enc_note",
-      author: null,
-    });
-
-    expect(result).toEqual({ id: "note_1", version: 1 });
-    expect(client.note.create.mutate).toHaveBeenCalledWith({
-      systemId: TEST_SYSTEM_ID,
-      encryptedData: "enc_note",
-      author: null,
-    });
-  });
-
-  it("update calls note.update with noteId", async () => {
-    const client = makeMockClient();
-    client.note.update.mutate.mockResolvedValue({ id: "note_1", version: 2 });
-    const api = createTRPCPersisterApi(client);
-
-    const result = await api.note.update(TEST_SYSTEM_ID, "note_1", {
-      encryptedData: "enc_note_v2",
-      version: 1,
-    });
-
-    expect(result).toEqual({ id: "note_1", version: 2 });
-    expect(client.note.update.mutate).toHaveBeenCalledWith({
-      systemId: TEST_SYSTEM_ID,
-      noteId: "note_1",
-      encryptedData: "enc_note_v2",
-      version: 1,
-    });
   });
 });
